LCORE-741: refactoring quota tables statements
